refactor(UserMissCheckingtest): 보너스 숫자 중복여부 LottoTest로 옮김
