+++ src/test/java/lotto/LottoTest.java
@@ -3,11 +3,16 @@ package lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class LottoTest {
+    private final List<Integer> dummyLotto = new ArrayList<>(List.of(1,3,4,6,12,15));
+    private final Lotto lotto = new Lotto(dummyLotto);
+
+
     @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
     @Test
     void createLottoByOverSize() {
@@ -23,5 +28,9 @@ class LottoTest {
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
-    // 아래에 추가 테스트 작성 가능
-}
+
+    @Test
+    public void 보너스숫자_선택한로또_중복여부(){
+        assertThatThrownBy(() -> lotto.addBonus(4))
+                .isInstanceOf(IllegalArgumentException.class);
+    }}
